Revert "Split metrics integration test"
This reverts commit 7f994e66
