--- tests/integration/__init__.py
@@ -1,2 +1,0 @@
-#  Copyright 2024 Canonical Ltd.
-#  See LICENSE file for licensing details.
